backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
ultralytics==8.1.9
easyocr==1.7.1
torch==2.1.2
torchvision
psutil
pytest
